stockrag/backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
google-genai==1.65.0
chromadb==1.5.2
sentence-transformers==5.2.3
pydantic==2.12.0
python-dotenv==1.0.0
httpx==0.28.1
